Attempt to use consistent format for lists of fragments
This is an attempt to be consistent in the formatting of lists of
sentence fragments. Apparently the django convention is to leave the
start of the fragment lowercase and avoid punctuation at the end. I'm
not clear on what to do when the list item begins with a fragment but
contains complete sentences following the fragment.